add --debug to powerlog

+++ pdf3json.py
@@ -198,21 +198,21 @@ for pdf_file in pdf_files:
             verbose_print("Merged OCR results saved to " + variable_str(os.path.join(json_folder, base_name + '.pdf.json')))
 
             # If not in debug mode, delete part files
-            #if not args.debug:
-            part_files = [f for f in os.listdir(divpdf_folder) if f.startswith(base_name + "_part")]
-            for part_file in part_files:
-                try:
-                    os.remove(os.path.join(divpdf_folder, part_file))
-                    warning_print(f"Deleted {part_file} in {divpdf_folder}")
-                except Exception as e:
-                    error_print(f"Failed to delete {part_file} in {divpdf_folder}. Reason: {e}")
-            part_files = [f for f in os.listdir(divjson_folder) if f.startswith(base_name + "_part")]
-            for part_file in part_files:
-                try:
-                    os.remove(os.path.join(divjson_folder, part_file))
-                    warning_print(f"Deleted {part_file} in {divjson_folder}")
-                except Exception as e:
-                    error_print(f"Failed to delete {part_file} in {divjson_folder}. Reason: {e}")
+            if not args.debug:
+                part_files = [f for f in os.listdir(divpdf_folder) if f.startswith(base_name + "_part")]
+                for part_file in part_files:
+                    try:
+                        os.remove(os.path.join(divpdf_folder, part_file))
+                        warning_print(f"Deleted {part_file} in {divpdf_folder}")
+                    except Exception as e:
+                        error_print(f"Failed to delete {part_file} in {divpdf_folder}. Reason: {e}")
+                part_files = [f for f in os.listdir(divjson_folder) if f.startswith(base_name + "_part")]
+                for part_file in part_files:
+                    try:
+                        os.remove(os.path.join(divjson_folder, part_file))
+                        warning_print(f"Deleted {part_file} in {divjson_folder}")
+                    except Exception as e:
+                        error_print(f"Failed to delete {part_file} in {divjson_folder}. Reason: {e}")
 
         else:
             process_pdf(pdf_file_path, document_intelligence_client, json_folder)
